+++ magda-correspondence-api/src/model.ts
@@ -2,4 +2,5 @@ export interface DatasetMessage {
     senderName: string;
     senderEmail: string;
     message: string;
+    note: string;
 }