+++ magda-correspondence-api/src/test/createApiRouter.spec.ts
@@ -218,19 +218,34 @@ describe("send dataset request mail", () => {
                 const args: Message = sendStub.firstCall.args[0];
 
                 expect(args.to).to.equal(DEFAULT_RECIPIENT);
+                expect(args.text).to.not.contain(" not a valid email address");
                 expect(args.subject).to.contain(
                     "for " + DEFAULT_DATASET_CONTACT_POINT
                 );
             });
         });
 
+        it("should find valid email address within contact point string", () => {
+            return sendQuestion({
+                contactPoint:
+                    "Commonwealth of Australia (Dept. of Code), sales@dept.gov.au"
+            }).then(() => {
+                const args: Message = sendStub.firstCall.args[0];
+
+                expect(args.to).to.equal("sales@dept.gov.au");
+                expect(args.text).to.not.contain(" not a valid email address");
+            });
+        });
+
         it("should fall back to default recipient if there's no contact point", () => {
             return sendQuestion({
-                contactPoint: undefined
+                contactPoint: undefined,
+                email: undefined
             }).then(() => {
                 const args: Message = sendStub.firstCall.args[0];
 
                 expect(args.to).to.equal(DEFAULT_RECIPIENT);
+                expect(args.text).to.contain(" not a valid email address");
             });
         });
 
@@ -241,6 +256,7 @@ describe("send dataset request mail", () => {
                 const args: Message = sendStub.firstCall.args[0];
 
                 expect(args.to).to.equal(DEFAULT_RECIPIENT);
+                expect(args.text).to.contain(" not a valid email address");
             });
         });
 
@@ -251,6 +267,7 @@ describe("send dataset request mail", () => {
                 const args: Message = sendStub.firstCall.args[0];
 
                 expect(args.to).to.equal(DEFAULT_RECIPIENT);
+                expect(args.text).to.contain(" not a valid email address");
             });
         });
 
@@ -262,6 +279,7 @@ describe("send dataset request mail", () => {
                 const args: Message = sendStub.firstCall.args[0];
 
                 expect(args.to).to.equal(DEFAULT_RECIPIENT);
+                expect(args.text).to.contain(" not a valid email address");
             });
         });
 
@@ -298,7 +316,7 @@ describe("send dataset request mail", () => {
     function stubGetRecordCall(overrideObject: {} = {}) {
         registryScope
             .get(
-                `/records/${ENCODED_DEFAULT_DATASET_ID}?aspect=dcat-dataset-strings&dereference=false`
+                `/records/${ENCODED_DEFAULT_DATASET_ID}?aspect=dcat-dataset-strings&optionalAspect=dataset-publisher&dereference=false`
             )
             .reply(200, {
                 id: DEFAULT_DATASET_ID,
@@ -308,6 +326,16 @@ describe("send dataset request mail", () => {
                         publisher: DEFAULT_DATASET_PUBLISHER,
                         contactPoint: DEFAULT_DATASET_CONTACT_POINT,
                         ...overrideObject
+                    },
+                    "dataset-publisher": {
+                        publisher: {
+                            aspects: {
+                                "organization-details": {
+                                    email: "contact@dept.gov.au",
+                                    ...overrideObject
+                                }
+                            }
+                        }
                     }
                 }
             });
@@ -337,7 +365,7 @@ describe("send dataset request mail", () => {
                 it("should return 404 if getting dataset from registry returns 404", () => {
                     registryScope
                         .get(
-                            `/records/${ENCODED_DEFAULT_DATASET_ID}?aspect=dcat-dataset-strings&dereference=false`
+                            `/records/${ENCODED_DEFAULT_DATASET_ID}?aspect=dcat-dataset-strings&optionalAspect=dataset-publisher&dereference=false`
                         )
                         .reply(404);
 
@@ -363,7 +391,7 @@ describe("send dataset request mail", () => {
                 it("should return 500 if getting dataset from registry returns 500", () => {
                     registryScope
                         .get(
-                            `/records/${ENCODED_DEFAULT_DATASET_ID}?aspect=dcat-dataset-strings&dereference=false`
+                            `/records/${ENCODED_DEFAULT_DATASET_ID}?aspect=dcat-dataset-strings&optionalAspect=dataset-publisher&dereference=false`
                         )
                         .reply(500);