+++ CHANGES.md
@@ -6,6 +6,10 @@ UI:
 -   Integrate sentry release notification and source map upload
 -   Added "role" to account page
 
+Correspondence:
+
+-   Broadened search for a valid dataset email contact address
+
 Others:
 
 -   Upgraded JDK version for magda-builder-scala to 8u201
@@ -19,7 +23,7 @@ Connectors:
 
 UI:
 
--   Upgraded the to create-react-app 2
+-   Upgraded to create-react-app 2
 -   Fixed javascript error when user clicks away from facet dropdowns
 -   Distribution pages: prepend file-specific icon to title
 -   Added a basic new dataset page